refactor(backend): split learning store hydrate concerns
